added limits

+++ app/src/main/java/com/example/tinkoffproject/model/data/dto/Wallet.kt
@@ -7,5 +7,5 @@ data class Wallet(
     val incomeAmount: Int,
     val expensesAmount: Int,
     val currency: Currency,
-    val limit: Long?
+    val limit: Int?
 )
